refactor(loop): extract write-guard subsystem to loop/write-guard.ts
